startup-intern-scraper/requirements.txt
beautifulsoup4>=4.12.0
fpdf2>=2.7.8
lxml>=4.9.0
pandas>=2.1.0
openpyxl>=3.1.0
plyer>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
